Switch to percy-client's batch resource upload.
